Fixed missing candidate members and updated with crawler changes
