+++ experiments/multichain/multichain_crawl_simple.conf
@@ -35,7 +35,7 @@ experiment_server_cmd = 'experiment_server.py'
 # remote_setup_cmd =
 #
 # Command used to start (the local part of) the experiment. This will be executed concurrently with its remote counterpart if set (see below)
-local_instance_cmd = "process_guard.py -c multichain_crawl_simple.py  -c multichain_crawl_simple.py -c multichain_crawl_simple.py -c multichain_crawl_simple.py  -c multichain_crawl_simple.py -c multichain_crawl_simple.py"
+local_instance_cmd = "process_guard.py -c multichain_client.py  -c multichain_client.py -c multichain_client.py -c multichain_client.py  -c multichain_client.py -c multichain_client.py"
 
 #
 # Command used to start (the remote part of) the experiment. This will be executed concurrently with its local counterpart if set (see above)
