--- experiments/multichain/multichain_crawler.scenario
@@ -1,9 +1,0 @@
-@0:0 set_master_member 3081a7301006072a8648ce3d020106052b81040027038192000406712a1e5381d030ae25697bd6626810af48ed5094e5faae18a8b91e482e60b8681b26bf6aca786ad81b207ffb5a755956b4d6353c9dc487eb1bfdd17b66bd0de444e01a375fcf3401177a7bb102da7c5f50ad9375fe4a1ae1baca9d47870f56c841169fc36c2b113bafbadacab9f2b3c7773b3d82a54cda8fbde41b9e0e571ba631997a8259cc78e7f262e933542b51
-@0:0 set_multichain_type MultiChainCommunityCrawler {1}
-@0:0 start_dispersy
-@0:1 online
-@0:5 reset_dispersy_statistics
-@0:10 annotate start-experiment-1-peer
-@0:3580 close
-@0:3590 stop_dispersy
-@0:3595 stop
