+++ experiments/multichain/multichain_standalone.scenario
@@ -6,18 +6,13 @@
 @0:1 online
 @0:5 reset_dispersy_statistics
 @0:10 annotate start-experiment-1-peer
-@0:15 introduce_candidates {1}
-@0:20 introduce_candidates {2}
-@0:25 introduce_candidates {3}
-@0:30 introduce_candidates {4}
-@0:35 introduce_candidates {5}
-@0:40 introduce_candidates {6}
-@0:80 request_signature 2 {1}
-@0:84 request_signature 1 {2}
-@0:94 request_signature 4 {1}
-@0:95 request_signature 1 {3}
-@0:104 request_signature 5 {1}
-@0:106 request_block 1 5 {6}
-@0:110 close
-@0:111 stop_dispersy
-@0:112 stop
+@0:15 introduce_candidates
+@0:20 request_signature 2 {1}
+@0:24 request_signature 1 {2}
+@0:34 request_signature 4 {1}
+@0:35 request_signature 1 {3}
+@0:44 request_signature 5 {1}
+@0:46 request_crawl 1 5 {6}
+@0:50 close
+@0:51 stop_dispersy
+@0:52 stop
